feat: Added todo visibility
Todo now have a a visibility field. This is used to determine if a user can see a todo
not owned by them. Aded the visibility field when creating a todo.
Fixed an issue where the todo item would clear while clicking on several
todo items. Now the debounce is global instead of per todo item. This
way you have time to click on several todo items before the debounce.